+++ Data/SBV/Examples/Existentials/CRCPolynomial.hs
@@ -71,7 +71,7 @@ genPoly hd = do res <- allSat $ do
                         -- the least significant bit must be set in the
                         -- polynomial, as all CRC polynomials have the "+1"
                         -- term in them set. This simplifies the query.
-                        return $ sbvTestBit p 0 &&& crcGood hd p s r
+                        return $ sTestBit p 0 &&& crcGood hd p s r
                 cnt <- displayModels disp res
                 putStrLn $ "Found: " ++ show cnt ++ " polynomail(s)."
         where disp :: Int -> (Bool, Word16) -> IO ()