+++ SBVUnitTest/TestSuite/Existentials/CRCPolynomial.hs
@@ -29,4 +29,4 @@ testSuite = mkTestSuite $ \goldCheck -> test [
                 r <- do rh <- forall "rh"
                         rl <- forall "rl"
                         return (rh, rl)
-                output $ sbvTestBit p 0 &&& crcGood 4 p s r
+                output $ sTestBit p 0 &&& crcGood 4 p s r